Move classifier into its own file to fix warning
Otherwise, cupbearer imports cupbearer.detectors.finetuning,
which imports the classifier, which before was in train_classifier.
So calling python -m train_classifier leads to an import order
warning.
